Refactor: Extract TableOfContents component and add to blog posts
